Simplified BaseIntegrationTest by removing the need to use Guice. Also added support for dropping all MongoDBs used for test to ensure isolation.
A major change that was required was setting forkMode=always in the pom.
This is needed to ensure that shutdown hooks are called between tests.
This should be revisted such that it is possible to perform this
without waiting for the JVM to exit. This could be done by having
a method exposed to the Main class.

One minor change was renaming the release2 dir to release3 to
better ensure test isolation.