Added project structure for etl component.
Encapsulated sub-module version references.

Isolated use of type-config Config object in the loader to just the integration tests and the ConfigModule. Created new ConfigModule in the loader since the configuration is simpler than in the server (which it previously used).

Package movements to align with the other components.

Added support for passing release name on the command line instead of specifying it in the mongo uri in the configuration file. This allows dynamic use of the service.

Added support for passing release to service in business method instead of constructor. This is more API friendly.

Major refactoring to remove dependency on TypeConfig. Harmonized configuration across etl sub-modules.

Formatting and spelling corrections.

Class renaming for alignment. Lombokifications.